requirements.txt
streamlit
PyMuPDF
nltk
spacy
torch
transformers
sentence-transformers
langchain
python-dotenv
faiss-cpu